Reorganize test data, make tests side-effect-free
